requirements.txt
groq>=0.4.0
feedparser>=6.0.10
httpx>=0.27.0
rich>=13.7.0
typer>=0.12.0
pyyaml>=6.0.1
python-dateutil>=2.9.0
plyer>=2.1.0
beautifulsoup4>=4.12.0
lxml>=5.0.0
